feat: move enum
